simplified project by extracting components
