+++ src/app/app.component.html
@@ -5,8 +5,21 @@
     </mat-grid-tile>
     <mat-grid-tile 
     [colspan] = "3"
-    [rowspan] = "1"><app-select-enemy></app-select-enemy>
-    <!-- <app-select-role></app-select-role> -->
+    [rowspan] = "1">
+
+    <form class="example-form">
+        <mat-form-field class="example-full-width">
+          <input matInput placeholder="Champion" aria-label="State" [matAutocomplete]="auto" [formControl]="stateCtrl" (change)="selectedEnemy($event.value)">
+          <mat-autocomplete #auto="matAutocomplete">
+            <mat-option *ngFor="let state of filteredStates | async" [value]="state.name">
+              <img class="example-option-img" aria-hidden [src]="state.icon" height="25">
+              <span>{{state.name}}</span> |
+              <small>Roles: {{state.roles}}</small>
+            </mat-option>
+          </mat-autocomplete>
+        </mat-form-field>
+      </form>
+      
     </mat-grid-tile>
     <mat-grid-tile 
     [colspan] = "1"
@@ -19,26 +32,47 @@
     <mat-grid-tile 
     [colspan] = "3"
     [rowspan] = "2">
-    <app-champ-profile style="width:100%;padding-top: 10%;"></app-champ-profile>
+
+    <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
+      
+        <!-- Position Column -->
+        <ng-container matColumnDef="position">
+          <th mat-header-cell *matHeaderCellDef> No. </th>
+          <td mat-cell *matCellDef="let element"> {{element.position}} </td>
+        </ng-container>
+      
+        <!-- Icon Column -->
+        <ng-container matColumnDef="icon">
+          <th mat-header-cell *matHeaderCellDef> Icon </th>
+          <td mat-cell *matCellDef="let element"> 
+            <img class="example-option-img" aria-hidden [src]="element.icon" height="25">  
+        </td>
+        </ng-container>
+      
+        <!-- Name Column -->
+        <ng-container matColumnDef="name">
+          <th mat-header-cell *matHeaderCellDef> Name </th>
+          <td mat-cell *matCellDef="let element"> {{element.name}} </td>
+        </ng-container>
+      
+        <!-- Tags Column -->
+        <ng-container matColumnDef="tags">
+          <th mat-header-cell *matHeaderCellDef> Tags </th>
+          <td mat-cell *matCellDef="let element"> {{element.tags}} </td>
+        </ng-container>
+      
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+      </table>
+
     </mat-grid-tile>
     <mat-grid-tile 
     [colspan] = "1"
     [rowspan] = "2">
     </mat-grid-tile>
-    <!-- <mat-grid-tile 
-    [colspan] = "1"
-    [rowspan] = "1">
-    </mat-grid-tile>
-    <mat-grid-tile 
-    [colspan] = "3"
-    [rowspan] = "1">
-    </mat-grid-tile>
-    <mat-grid-tile 
-    [colspan] = "1"
-    [rowspan] = "1">
-    </mat-grid-tile> -->
     <mat-grid-tile 
     [colspan] = "5"
-    [rowspan] = 1> <button mat-raised-button (click)="openDialog()" color="primary">Submit</button>
+    [rowspan] = 1> 
+    <!-- replaced submit button by enter / select in search -- REMEBER TO USE openDialog() function -->
     </mat-grid-tile>
  </mat-grid-list>
