+++ src/app/app.component.ts
@@ -1,7 +1,30 @@
 import {Component, Inject, ViewChild} from '@angular/core';
 import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
-import {SelectEnemyComponent} from './select-enemy/select-enemy.component';
 import data from './../assets/test.json';
+import {FormControl} from '@angular/forms';
+import {Observable} from 'rxjs';
+import {map, startWith} from 'rxjs/operators';
+
+
+export interface PeriodicElement {
+  icon: string;
+  position: number;
+  name: String;
+  tags: string;
+}
+
+const ELEMENT_DATA: PeriodicElement[] = [
+  {position: 1, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Fiora.png', name: "Fiora", tags: 'Physical Damage, Fighter'},
+  {position: 2, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Fizz.png', name: "Fizz", tags: 'Fighter", Magical Damage'},
+  {position: 3, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Ivern.png', name: "Ivern", tags: 'Support", Magical Damage'},
+  {position: 4, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Fiora.png', name: "Fiora", tags: 'Physical Damage, Fighter'},
+  {position: 5, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Fizz.png', name: "Fizz", tags: 'Fighter", Magical Damage'},
+  {position: 6, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Ivern.png', name: "Ivern", tags: 'Support", Magical Damage'},
+  {position: 7, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Fiora.png', name: "Fiora", tags: 'Physical Damage, Fighter'},
+  {position: 8, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Fizz.png', name: "Fizz", tags: 'Fighter", Magical Damage'},
+  {position: 9, icon: 'http://ddragon.leagueoflegends.com/cdn/8.24.1/img/champion/Ivern.png', name: "Ivern", tags: 'Support", Magical Damage'}
+];
+
 
 export interface DialogData {
   enemyChamp: string;
@@ -19,15 +42,31 @@ export interface DialogData {
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
+  stateCtrl = new FormControl();
+  filteredStates: Observable<any>;
   champions = data;
+  enemyChamp: string;
+
+  constructor(public dialog: MatDialog) {
+    this.filteredStates = this.stateCtrl.valueChanges
+      .pipe(
+        startWith(''),
+        map(state => state ? this._filterStates(state) : this.champions.slice())
+      );
+  }
+
+  private _filterStates(value: string) {
+    const filterValue = value.toLowerCase();
+    this.enemyChamp = value;
+    return this.champions.filter(state => state.name.toLowerCase().indexOf(filterValue) === 0);
+  }
+
   selectedEnemy : string;
-  @ViewChild(SelectEnemyComponent, {static: true}) enemy: SelectEnemyComponent;
+  displayedColumns: string[] = ['position', 'icon', 'name', 'tags'];
+  dataSource = ELEMENT_DATA;
 
-
-  constructor(public dialog: MatDialog) {}
-
-  openDialog(): void {
-    this.selectedEnemy = this.enemy.enemy;
+  openDialog(): void { // messed up, fix 
+    // this.selectedEnemy = this.enemy;
     let counters: string[];
     let enemyChampTips: string[];
     let bestPickName: string;
