+++ src/app/app.module.ts
@@ -5,27 +5,19 @@ import { AppComponent } from './app.component';
 
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {MatGridListModule} from '@angular/material/grid-list';
-import { SelectRoleComponent } from './select-role/select-role.component';
 import {MatSelectModule} from '@angular/material/select';
-import { SelectPowerComponent } from './select-power/select-power.component';
 import {MatCheckboxModule} from '@angular/material/checkbox';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {MatAutocompleteModule, MatInputModule} from '@angular/material';
-import { SelectEnemyComponent } from './select-enemy/select-enemy.component';
 import {MatDialogModule} from '@angular/material/dialog';
 import {SubmitDialog} from './app.component';
-import { ChampProfileComponent } from './champ-profile/champ-profile.component';
 import {MatTableModule} from '@angular/material/table';
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    SelectRoleComponent,
-    SelectPowerComponent,
-    SelectEnemyComponent,
-    SubmitDialog,
-    ChampProfileComponent
+    SubmitDialog
   ],
   imports: [
     BrowserModule,
