--- src/app/champ-profile/champ-profile.component.css
@@ -1,3 +1,0 @@
-table {
-    width: 100%;
-  }
