--- src/app/champ-profile/champ-profile.component.html
@@ -1,34 +1,0 @@
-<table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
-
-    <!--- Note that these columns can be defined in any order.
-          The actual rendered columns are set as a property on the row definition" -->
-  
-    <!-- Position Column -->
-    <ng-container matColumnDef="position">
-      <th mat-header-cell *matHeaderCellDef> No. </th>
-      <td mat-cell *matCellDef="let element"> {{element.position}} </td>
-    </ng-container>
-  
-    <!-- Icon Column -->
-    <ng-container matColumnDef="icon">
-      <th mat-header-cell *matHeaderCellDef> Icon </th>
-      <td mat-cell *matCellDef="let element"> 
-        <img class="example-option-img" aria-hidden [src]="element.icon" height="25">  
-    </td>
-    </ng-container>
-  
-    <!-- Name Column -->
-    <ng-container matColumnDef="name">
-      <th mat-header-cell *matHeaderCellDef> Name </th>
-      <td mat-cell *matCellDef="let element"> {{element.name}} </td>
-    </ng-container>
-  
-    <!-- Tags Column -->
-    <ng-container matColumnDef="tags">
-      <th mat-header-cell *matHeaderCellDef> Tags </th>
-      <td mat-cell *matCellDef="let element"> {{element.tags}} </td>
-    </ng-container>
-  
-    <!-- <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr> -->
-    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-  </table>
