--- src/app/select-enemy/select-enemy.component.html
@@ -1,15 +1,0 @@
-<h4>Who are you playing against?</h4>
-<form class="example-form">
-        <mat-form-field class="example-full-width">
-          <input matInput placeholder="Champion" aria-label="State" [matAutocomplete]="auto" [formControl]="stateCtrl" (change)="selectedEnemy($event.value)">
-          <mat-autocomplete #auto="matAutocomplete">
-            <mat-option *ngFor="let state of filteredStates | async" [value]="state.name">
-              <img class="example-option-img" aria-hidden [src]="state.icon" height="25">
-              <span>{{state.name}}</span> |
-              <small>Roles: {{state.roles}}</small>
-            </mat-option>
-          </mat-autocomplete>
-        </mat-form-field>
-      </form>
-
-
