--- src/app/select-enemy/select-enemy.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SelectEnemyComponent } from './select-enemy.component';
-
-describe('SelectEnemyComponent', () => {
-  let component: SelectEnemyComponent;
-  let fixture: ComponentFixture<SelectEnemyComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ SelectEnemyComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(SelectEnemyComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
