--- src/app/select-power/select-power.component.css
@@ -1,21 +1,0 @@
-.example-h2 {
-    margin: 10px;
-  }
-  
-  .example-section {
-    display: flex;
-    align-content: center;
-    align-items: center;
-    height: 60px;
-  }
-  
-  .example-margin {
-    margin: 0 10px;
-  }
-
-  .container {
-    width: inherit;
-  }
-
-  .mat-checkbox-layout { white-space: normal !important; }
-
