--- src/app/select-power/select-power.component.html
@@ -1,12 +1,0 @@
-<div class="container">
-  <h4>Pick your power:</h4>
-
-  <section class="example-section">
-      <mat-checkbox class="example-margin" [(ngModel)]="ad">AD</mat-checkbox>
-      <mat-checkbox class="example-margin" [(ngModel)]="ap">AP</mat-checkbox>
-  </section>
-  <section>
-      <mat-checkbox class="example-margin" [(ngModel)]="tank">Tank</mat-checkbox>
-      <mat-checkbox class="example-margin" [(ngModel)]="utility">Utility</mat-checkbox>
-  </section>
-</div>
