--- src/app/select-role/select-role.component.html
@@ -1,9 +1,0 @@
-<h4>Pick your lane:</h4>
-<mat-form-field>
-  <mat-label>Lane</mat-label>
-  <mat-select>
-    <mat-option *ngFor="let lane of lanes" [value]="lane.value">
-      {{lane.viewValue}}
-    </mat-option>
-  </mat-select>
-</mat-form-field>
